backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
python-multipart==0.0.9
pillow==10.2.0
numpy==1.26.4
tensorflow==2.15.0
opencv-python==4.9.0.80
pydantic==2.6.3
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.2